Modified   SourceCode/Engine/Engine/ResourceSystem/Private/ImExporter.cpp Modified   SourceCode/Engine/Engine/ResourceSystem/Private/ResourceFactory.cpp Modified   SourceCode/Engine/Engine/ResourceSystem/Private/ResourceHolder.cpp Modified   SourceCode/Engine/Engine/Utility/AssetParser/InternalModelParser.cpp Modified   SourceCode/Engine/Engine/Utility/Lexer/Token.cpp Modified   SourceCode/Engine/Engine/Utility/Lexer/Tokenizer.cpp Modified   SourceCode/Engine/Engine/Utility/YAML/YAMLCodeParser.cpp Modified   SourceCode/Engine/Engine/Utility/YAML/YAMLParser.cpp Modified   SourceCode/Engine/Engine/Containers/AnyDataType.h Modified   SourceCode/Engine/Engine/Containers/DynamicString.h Modified   SourceCode/Engine/Engine/ResourceSystem/Private/ImExporter.h Modified   SourceCode/Engine/Engine/ResourceSystem/Private/ResourceFactory.h Modified   SourceCode/Engine/Engine/Utility/AssetParser/InternalModelParser.h Modified   SourceCode/Engine/Engine/Utility/Lexer/Token.h Modified   SourceCode/Engine/Engine/Utility/YAML/YAMLCodeParser.h

+++ SourceCode/Engine/Engine/Utility/Lexer/Tokenizer.cpp
@@ -96,9 +96,9 @@ namespace Engine
 					if (isFloat)
 						Token.SetConstantFloat32(StringUtility::ToFloat32(Token.GetIdentifier()));
 					else if (isHex)
-						Token.SetConstantInt32(StringUtility::ToFloat32(Token.GetIdentifier()));
+						Token.SetConstantInt32(StringUtility::ToInt32(Token.GetIdentifier()));
 					else
-						Token.SetConstantInt32(StringUtility::ToFloat32(Token.GetIdentifier()));
+						Token.SetConstantInt32(StringUtility::ToInt32(Token.GetIdentifier()));
 
 					return true;
 				}
